requirements.txt
streamlit
fpdf
psycopg2-binary
